+++ src/components/ChartTasks/ChartTasks.module.scss
@@ -2,7 +2,7 @@
 @use '../../styles/abstracts/variables' as *;
 @use '../../styles/abstracts/breakpoints' as *;
 
-.table_tasks {
+.chart_tasks {
   width: 390px;
   min-height: calc(100vh - $header-height - $global-indent - 16px - $global-indent * 0.5);
   overflow: hidden;
