+++ src/components/ChartTimeline/ChartTimeline.module.scss
@@ -2,7 +2,7 @@
 @use '../../styles/abstracts/variables' as *;
 @use '../../styles/abstracts/breakpoints' as *;
 
-.table_timeline {
+.chart_timeline {
   position: relative;
 
   flex-grow: 1;
