+++ src/modules/Router/routes.tsx
@@ -1,9 +1,9 @@
 import { Navigate } from 'react-router-dom';
 import { NotFoundPage } from 'pages/NotFoundPage/NotFoundPage';
-import { TablePage } from '../../pages/TablePage/TablePage';
+import { ChartPage } from '../../pages/ChartPage/ChartPage';
 
 export const ROUTES_LIST = [
-  { path: '/', component: <TablePage /> },
+  { path: '/', component: <ChartPage /> },
   { path: '/404', component: <NotFoundPage /> },
   { path: '*', component: <Navigate to="/404" /> },
 ];
