+++ src/store/utils.ts
@@ -29,3 +29,30 @@ export const toggleOpenClose = (elem: ChartTask, fId: number) => {
   }
   return false;
 };
+
+//
+//
+// export function Family({ familyTree }) {
+//   const [isVisible, setIsVisible] = useState(false);
+//   const expand = () => {
+//     setIsVisible(!isVisible);
+//   };
+//   return (
+//     <div style={{ paddingLeft: 10 }}>
+//   <span onClick={expand}>{familyTree.name}</span>
+//   {isVisible ? (
+//     familyTree?.children?.map((child) => {
+//       return (
+//         <div style={{ paddingLeft: 10 }}>
+//       <Family familyTree={child} />
+//       </div>
+//     );
+//     })
+//   ) : (
+//     <></>
+//   )}
+//   </div>
+// );
+// }
+//
+// <Family familyTree={familyTree} />
